Gpg integration (#16)
Integrated GPG for enhanced security—private keys are now encrypted with a password.
Added keyring-manager for tracking, hiding/revealing, and exporting keys.
Improved dependency checking mechanism—missing software is now automatically installed if possible.
Stake key deregistration now automatically withdraws rewards.
Public keys or wallet addresses are automatically derived if they are missing.
CFD now automatically tracks and manages the state of keys in use.
Fixed several minor bugs.
Updated default software versions in the configuration.

+++ scripts/software-tools.sh
@@ -265,9 +265,15 @@ function software_deploy(){
     
     if ! [ "$SF_LOCAL_META" == null ]; then
         local DESIRED_SF_VERSION=$(get-sf-version $SF_NAME)
-        DESIRED_SF_VERSION_BASE=$(echo $DESIRED_SF_VERSION | awk -F'-' '{print $1}')
-        DESIRED_SF_VERSION_SUFFIX=$(echo "$VERSION" | awk -F'-' '{if (NF>1) print "-"$2; else print ""}')
-        
+        if [ "$(echo "$DESIRED_SF_VERSION" | awk -F'-' '{print NF-1}')" -eq 1 ]; then
+            DESIRED_SF_VERSION_BASE=$(echo "$DESIRED_SF_VERSION" | awk -F'-' '{print $1}')
+            DESIRED_SF_VERSION_SUFFIX=$(echo "$DESIRED_SF_VERSION" | awk -F'-' '{print "-"$2}')
+        else
+            DESIRED_SF_VERSION_BASE="$DESIRED_SF_VERSION"
+            DESIRED_SF_VERSION_SUFFIX=""
+        fi
+
+
         local SF_GLOBAL_DIR=$CARDANO_SOFTWARE_DIR/$SF_NAME
         local SF_LOCAL_DIR=$SF_GLOBAL_DIR/$DESIRED_SF_VERSION
         local SF_BIN_DIR=$SF_LOCAL_DIR/bin
@@ -286,7 +292,7 @@ function software_deploy(){
             
             if [ "$VERBOSITY" != "silent" ]; then
                 echo ""
-                echo "$SF_NAME ver: $DESIRED_SF_VERSION  not found"   
+                echo "$SF_NAME ver: $DESIRED_SF_VERSION not found"
                 echo "Installing..."   
                 echo ""
             fi 
@@ -390,7 +396,11 @@ function recursive-config-linking {
         ln -fns "$NEW_USER_CONF_SUBJECT" "$LINK_CONF_SUBJECT"         
         
         if [ ! -e "$NEW_USER_CONF_SUBJECT" ]; then
-            cp -r "$OLD_USER_CONF_SUBJECT" "$NEW_USER_CONF_SUBJECT"
+            if [ -e "$OLD_USER_CONF_SUBJECT" ]; then
+                cp -r "$OLD_USER_CONF_SUBJECT" "$NEW_USER_CONF_SUBJECT"
+            else
+                cp -r "$NEW_DEF_CONF_SUBJECT" "$NEW_USER_CONF_SUBJECT"
+            fi
         fi
 
         if [ -f "$SUBJECT" ]; then
